refactor: refactor codespace to match project convention
