+++ package.json
@@ -18,6 +18,7 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-hook-form": "^7.52.1",
+    "react-router-dom": "^6.25.1",
     "sass": "^1.77.8",
     "yup": "^1.4.0"
   },
