+++ src/App.tsx
@@ -1,8 +1,5 @@
-// import { useState } from 'react';
-// import reactLogo from './assets/react.svg';
-// import viteLogo from '/vite.svg';
+import LoginForm from './modules/auth/components/LoginForm';
 import './App.css';
-import LoginForm from './features/auth/components/login-form';
 
 function App() {
   // const [count, setCount] = useState(0);
