+++ src/modules/auth/assets/company-logos.png
