+++ src/modules/auth/assets/graduation.png
